+++ poo-2022-p2-musicos-igorney/src/main/java/org/ufabc/musico/instrumento/corda/Violao.java
@@ -4,9 +4,9 @@ import org.ufabc.musico.instrumento.Instrumento;
 import org.ufabc.musico.instrumento.interfaces.Distorcao;
 
 public class Violao extends Instrumento implements Distorcao {
-    private final Long cordas;
+    private final int cordas;
 
-    public Violao(Long cordas) {
+    public Violao(int cordas) {
         this.cordas = cordas;
     }
 
@@ -15,7 +15,7 @@ public class Violao extends Instrumento implements Distorcao {
         return emitirSomDistorcido();
     }
 
-    public Long getCordas() {
+    public int getCordas() {
         return cordas;
     }
